+++ client/http-client/src/client.rs
@@ -32,11 +32,10 @@ use crate::types::{ErrorResponse, Id, NotificationSer, RequestSer, Response};
 use async_trait::async_trait;
 use hyper::http::HeaderMap;
 use jsonrpsee_core::client::{CertificateStore, ClientT, IdKind, RequestIdManager, Subscription, SubscriptionClientT};
+use jsonrpsee_core::params::{BatchRequestBuilder, ToRpcParams};
 use jsonrpsee_core::tracing::RpcTracing;
 use jsonrpsee_core::{Error, TEN_MB_SIZE_BYTES};
 use jsonrpsee_types::error::CallError;
-use jsonrpsee_types::params::ToRpcParams;
-use jsonrpsee_types::BatchRequestBuilder;
 use rustc_hash::FxHashMap;
 use serde::de::DeserializeOwned;
 use tracing_futures::Instrument;
