+++ proc-macros/tests/ui/correct/basic.rs
@@ -5,7 +5,8 @@ use std::net::SocketAddr;
 use jsonrpsee::core::{async_trait, client::ClientT, RpcResult};
 use jsonrpsee::proc_macros::rpc;
 use jsonrpsee::rpc_params;
-use jsonrpsee::types::{SubscriptionResult, ArrayParams, EmptyParams};
+use jsonrpsee::core::params::{ArrayParams, EmptyParams};
+use jsonrpsee::types::SubscriptionResult;
 use jsonrpsee::ws_client::*;
 use jsonrpsee::ws_server::{SubscriptionSink, WsServerBuilder};
