+++ tests/tests/integration_tests.rs
@@ -35,12 +35,12 @@ use helpers::{http_server, http_server_with_access_control, websocket_server, we
 use hyper::http::HeaderValue;
 use jsonrpsee::core::client::{ClientT, IdKind, Subscription, SubscriptionClientT};
 use jsonrpsee::core::error::SubscriptionClosed;
+use jsonrpsee::core::params::{ArrayParams, BatchRequestBuilder, EmptyParams};
 use jsonrpsee::core::{Error, JsonValue};
 use jsonrpsee::http_client::HttpClientBuilder;
 use jsonrpsee::http_server::AccessControlBuilder;
 use jsonrpsee::rpc_params;
 use jsonrpsee::types::error::ErrorObject;
-use jsonrpsee::types::{ArrayParams, BatchRequestBuilder, EmptyParams};
 use jsonrpsee::ws_client::WsClientBuilder;
 use tokio::time::interval;
 use tokio_stream::wrappers::IntervalStream;
