+++ types/src/lib.rs
@@ -43,10 +43,7 @@ pub mod response;
 pub mod error;
 
 pub use error::{ErrorObject, ErrorObjectOwned, ErrorResponse, SubscriptionEmptyError, SubscriptionResult};
-pub use params::{
-	ArrayParams, ArrayParamsBuilder, BatchRequestBuilder, EmptyParams, Id, ObjectParams, ObjectParamsBuilder, Params,
-	ParamsSequence, ParamsSer, SubscriptionId, ToRpcParams, TwoPointZero,
-};
+pub use params::{Id, Params, ParamsSequence, ParamsSer, SubscriptionId, TwoPointZero};
 pub use request::{InvalidRequest, Notification, NotificationSer, Request, RequestSer};
 pub use response::{Response, SubscriptionPayload, SubscriptionResponse};
