+++ types/src/params.rs
@@ -36,15 +36,7 @@ use beef::Cow;
 use serde::de::{self, Deserializer, Unexpected, Visitor};
 use serde::ser::Serializer;
 use serde::{Deserialize, Serialize};
-use serde_json::value::RawValue;
 use serde_json::Value as JsonValue;
-
-#[doc(hidden)]
-pub mod __reexports {
-	pub use crate::params::ArrayParams;
-	pub use crate::params::ArrayParamsBuilder;
-	pub use crate::params::ToRpcParams;
-}
 
 /// JSON-RPC v2 marker type.
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
@@ -413,275 +405,6 @@ impl<'a> Id<'a> {
 			Id::Number(num) => Id::Number(num),
 			Id::Str(s) => Id::Str(Cow::owned(s.into_owned())),
 		}
-	}
-}
-
-/// Helper module for building parameters.
-mod params_builder {
-	use serde::Serialize;
-
-	/// Initial number of bytes for a parameter length.
-	const PARAM_BYTES_CAPACITY: usize = 128;
-
-	/// Generic parameter builder that serializes parameters to bytes.
-	/// This produces a JSON compatible String.
-	///
-	/// The implementation relies on `Vec<u8>` to hold the serialized
-	/// parameters in memory for the following reasons:
-	///   1. Other serialization methods than `serde_json::to_writer` would internally
-	///      have an extra heap allocation for temporarily holding the value in memory.
-	///   2. `io::Write` is not implemented for `String` required for serialization.
-	#[derive(Debug)]
-	pub struct ParamsBuilder {
-		bytes: Vec<u8>,
-		end: char,
-	}
-
-	impl ParamsBuilder {
-		/// Construct a new [`ParamsBuilder`] with custom start and end tokens.
-		/// The inserted values are wrapped by the _start_ and _end_ tokens.
-		fn new(start: char, end: char) -> Self {
-			let mut bytes = Vec::with_capacity(PARAM_BYTES_CAPACITY);
-			bytes.push(start as u8);
-			ParamsBuilder { bytes, end }
-		}
-
-		/// Construct a new [`ParamsBuilder`] for positional parameters equivalent to a JSON array object.
-		pub fn positional() -> Self {
-			Self::new('[', ']')
-		}
-
-		/// Construct a new [`ParamsBuilder`] for named parameters equivalent to a JSON map object.
-		pub fn named() -> Self {
-			Self::new('{', '}')
-		}
-
-		/// Insert a named value (key, value) pair into the builder.
-		/// The _name_ and _value_ are delimited by the `:` token.
-		pub fn insert_named<P: Serialize>(&mut self, name: &str, value: P) -> Result<(), serde_json::Error> {
-			serde_json::to_writer(&mut self.bytes, name)?;
-			self.bytes.push(b':');
-			serde_json::to_writer(&mut self.bytes, &value)?;
-			self.bytes.push(b',');
-
-			Ok(())
-		}
-
-		/// Insert a plain value into the builder.
-		pub fn insert<P: Serialize>(&mut self, value: P) -> Result<(), serde_json::Error> {
-			serde_json::to_writer(&mut self.bytes, &value)?;
-			self.bytes.push(b',');
-
-			Ok(())
-		}
-
-		/// Finish the building process and return a JSON compatible string.
-		pub fn build(mut self) -> String {
-			let idx = self.bytes.len() - 1;
-			if self.bytes[idx] == b',' {
-				self.bytes[idx] = self.end as u8;
-			} else {
-				self.bytes.push(self.end as u8);
-			}
-
-			// Safety: This is safe because JSON does not emit invalid UTF-8.
-			unsafe { String::from_utf8_unchecked(self.bytes) }
-		}
-	}
-}
-
-/// Parameter builder that serializes named value parameters to a JSON compatible string.
-/// This is the equivalent of a JSON Map object `{ key: value }`.
-///
-/// # Examples
-///
-/// ```rust
-///
-/// use jsonrpsee_types::ObjectParamsBuilder;
-///
-/// let mut builder = ObjectParamsBuilder::new();
-/// builder.insert("param1", 1);
-/// builder.insert("param2", "abc");
-/// let params = builder.build();
-///
-/// // Use RPC parameters...
-/// ```
-#[derive(Debug)]
-pub struct ObjectParamsBuilder(params_builder::ParamsBuilder);
-
-impl ObjectParamsBuilder {
-	/// Construct a new [`ObjectParamsBuilder`].
-	pub fn new() -> Self {
-		Self::default()
-	}
-
-	/// Insert a named value (key, value) pair into the builder.
-	/// The _name_ and _value_ are delimited by the `:` token.
-	pub fn insert<P: Serialize>(&mut self, name: &str, value: P) -> Result<(), serde_json::Error> {
-		self.0.insert_named(name, value)
-	}
-
-	/// Finish the building process and return a JSON compatible string.
-	pub fn build(self) -> ObjectParams {
-		ObjectParams(self.0.build())
-	}
-}
-
-impl Default for ObjectParamsBuilder {
-	fn default() -> Self {
-		Self(params_builder::ParamsBuilder::named())
-	}
-}
-
-/// Object RPC parameters stored as a JSON Map object `{ key: value }`.
-#[derive(Clone, Debug)]
-pub struct ObjectParams(String);
-
-impl ToRpcParams for ObjectParams {
-	fn to_rpc_params(self) -> Result<Option<Box<RawValue>>, serde_json::Error> {
-		RawValue::from_string(self.0).map(Some)
-	}
-}
-
-/// Parameter builder that serializes plain value parameters to a JSON compatible string.
-/// This is the equivalent of a JSON Array object `[ value0, value1, .., valueN ]`.
-///
-/// # Examples
-///
-/// ```rust
-///
-/// use jsonrpsee_types::ArrayParamsBuilder;
-///
-/// let mut builder = ArrayParamsBuilder::new();
-/// builder.insert("param1");
-/// builder.insert(1);
-/// let params = builder.build();
-///
-/// // Use RPC parameters...
-/// ```
-#[derive(Debug)]
-pub struct ArrayParamsBuilder(params_builder::ParamsBuilder);
-
-impl ArrayParamsBuilder {
-	/// Construct a new [`ArrayParamsBuilder`].
-	pub fn new() -> Self {
-		Self::default()
-	}
-
-	/// Insert a plain value into the builder.
-	pub fn insert<P: Serialize>(&mut self, value: P) -> Result<(), serde_json::Error> {
-		self.0.insert(value)
-	}
-
-	/// Finish the building process and return a JSON compatible string.
-	pub fn build(self) -> ArrayParams {
-		ArrayParams(self.0.build())
-	}
-}
-
-impl Default for ArrayParamsBuilder {
-	fn default() -> Self {
-		Self(params_builder::ParamsBuilder::positional())
-	}
-}
-
-/// Array RPC parameters stored as a JSON Array object `[ value0, value1, .., valueN ]`.
-#[derive(Clone, Debug)]
-pub struct ArrayParams(String);
-
-impl ToRpcParams for ArrayParams {
-	fn to_rpc_params(self) -> Result<Option<Box<RawValue>>, serde_json::Error> {
-		RawValue::from_string(self.0).map(Some)
-	}
-}
-
-/// Marker trait for types that can be serialized as JSON compatible strings.
-///
-/// This trait ensures the correctness of the RPC parameters.
-///
-/// # Note
-///
-/// Please consider using the [`ArrayParamsBuilder`] and [`ObjectParamsBuilder`] than
-/// implementing this trait.
-///
-/// # Examples
-///
-/// - Implementation for hard-coded strings
-///
-/// ```rust
-///
-/// use jsonrpsee_types::ToRpcParams;
-/// use serde_json::value::RawValue;
-///
-/// struct ManualParam;
-///
-/// impl ToRpcParams for ManualParam {
-///     fn to_rpc_params(self) -> Result<Option<Box<RawValue>>, serde_json::Error> {
-///         // Manually define a valid JSONRPC parameter.
-///         RawValue::from_string("[1, \"2\", 3]".to_string()).map(Some)
-///     }
-/// }
-/// ```
-///
-/// - Implementation for JSON serializable structures
-///
-/// ```rust
-/// use jsonrpsee_types::ToRpcParams;
-/// use serde_json::value::RawValue;
-/// use serde::Serialize;
-///
-/// #[derive(Serialize)]
-/// struct SerParam {
-///     param_1: u8,
-///     param_2: String,
-/// };
-///
-/// impl ToRpcParams for SerParam {
-///     fn to_rpc_params(self) -> Result<Option<Box<RawValue>>, serde_json::Error> {
-///         let s = String::from_utf8(serde_json::to_vec(&self)?).expect("Valid UTF8 format");
-///         RawValue::from_string(s).map(Some)
-///     }
-/// }
-/// ```
-pub trait ToRpcParams {
-	/// Consume and serialize the type as a JSON raw value.
-	fn to_rpc_params(self) -> Result<Option<Box<RawValue>>, serde_json::Error>;
-}
-
-/// Initial number of parameters in a batch request.
-const BATCH_PARAMS_NUM_CAPACITY: usize = 4;
-
-/// Request builder that serializes RPC parameters to construct a valid batch parameter.
-/// This is the equivalent of chaining multiple RPC requests.
-#[derive(Clone, Debug, Default)]
-pub struct BatchRequestBuilder<'a>(Vec<(&'a str, Option<Box<RawValue>>)>);
-
-impl<'a> BatchRequestBuilder<'a> {
-	/// Construct a new [`BatchRequestBuilder`].
-	pub fn new() -> Self {
-		Self(Vec::with_capacity(BATCH_PARAMS_NUM_CAPACITY))
-	}
-
-	/// Inserts the RPC method with provided parameters into the builder.
-	pub fn insert<Params: ToRpcParams>(&mut self, method: &'a str, value: Params) -> Result<(), serde_json::Error> {
-		self.0.push((method, value.to_rpc_params()?));
-		Ok(())
-	}
-
-	/// Finish the building process and return a valid batch parameter.
-	pub fn build(self) -> Vec<(&'a str, Option<Box<RawValue>>)> {
-		self.0
-	}
-}
-
-/// Empty RPC parameters that perform no allocation.
-#[derive(Clone, Debug)]
-pub struct EmptyParams;
-
-/// Custom implementation for empty RPC parameters.
-impl ToRpcParams for EmptyParams {
-	fn to_rpc_params(self) -> Result<Option<Box<RawValue>>, serde_json::Error> {
-		Ok(None)
 	}
 }
